refactor: ValueProvider 제거 및 패키지 이동
