--- src/main/kotlin/study/racing/SystemValueProvider.kt
@@ -1,5 +1,0 @@
-package study.racing
-
-object SystemValueProvider : ValueProvider<String?> {
-    override fun getValue(): String? = readLine()
-}
