--- src/main/kotlin/study/racing/ValueProvider.kt
@@ -1,5 +1,0 @@
-package study.racing
-
-interface ValueProvider<T> {
-    fun getValue(): T
-}
