+++ src/main/kotlin/study/racing/data/PowerSource.kt
@@ -1,4 +1,4 @@
-package study.racing
+package study.racing.data
 
 interface PowerSource {
     fun getPower(): Int
