+++ src/main/kotlin/study/racing/data/RandomPowerSource.kt
@@ -1,4 +1,4 @@
-package study.racing
+package study.racing.data
 
 class RandomPowerSource : PowerSource {
