+++ src/main/kotlin/study/racing/domain/repository/RacingCarRepository.kt
@@ -1,7 +1,5 @@
 package study.racing.domain.repository
 
-import study.racing.PowerSource
-
 interface RacingCarRepository {
     fun getPower(): Int
     fun canMoveForward(power: Int): Boolean
