+++ src/main/kotlin/study/racing/view/ResultView.kt
@@ -1,4 +1,4 @@
-package study.racing.ui
+package study.racing.view
 
 import study.racing.domain.entity.Race
