+++ src/test/kotlin/study/racing/FakePowerSource.kt
@@ -1,5 +1,7 @@
 package study.racing
 
+import study.racing.data.PowerSource
+
 class FakePowerSource(
     private val power: Int
 ) : PowerSource {
