+++ src/test/kotlin/study/racing/RaceTest.kt
@@ -1,6 +1,7 @@
 package study.racing
 
 import org.junit.jupiter.api.Test
+import study.racing.data.RandomPowerSource
 import study.racing.domain.entity.Race
 import study.racing.domain.entity.RacingCar
 import study.racing.domain.repository.RacingCarRepositoryImpl
