+++ src/test/kotlin/study/racing/RacingCarRepositoryTest.kt
@@ -2,6 +2,7 @@ package study.racing
 
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import study.racing.data.RandomPowerSource
 import study.racing.domain.repository.RacingCarRepositoryImpl
 
 class RacingCarRepositoryTest {
